Average Color of a Book Analysis
Takes a .txt file of a book and "reads" it, noting all usages of color in the book. Averages (based on the frequency and RGB values) all mentions of different colors in the book in order to come up with the "average" color of the book, which it displays.

Can also analyze many different books of one genre at a time.